fix: move TextEditorToolbar as a separate component
