+++ packages/picasso-lab/src/TextEditor/TextEditor.tsx
@@ -4,13 +4,13 @@ import { BaseProps } from '@toptal/picasso-shared'
 import cx from 'classnames'
 import { Container } from '@toptal/picasso'
 
+import QuillEditor from '../QuillEditor'
+import Toolbar from '../TextEditorToolbar'
 import styles from './styles'
-import Toolbar from './TextEditorToolbar'
 import { TextEditorChangeHandler } from './types'
 import { actions as toolbarActions } from './store/toolbar'
 import useTextEditorState from './hooks/useTextEditorState'
 import useHasFocus from './hooks/useHasFocus'
-import QuillEditor from '../QuillEditor'
 import useOnSelectionChange from './hooks/useOnSelectionChange'
 import { ToolbarHandlers } from './store/toolbar/types'
