+++ packages/picasso-lab/src/TextEditorToolbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TextEditorToolbar'
